+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed the execution time of the asset profile data gathering to every Sunday at lunch time
+- Improved the activities import by providing asset profile details
 - Upgraded `@codewithdan/observable-store` from version `2.2.11` to `2.2.15`
 - Upgraded `bull` from version `4.8.5` to `4.10.2`
 - Upgraded `countup.js` from version `2.0.7` to `2.3.2`